+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [1.18.3](https://github.com/adobe/spacecat-audit-worker/compare/v1.18.2...v1.18.3) (2024-06-18)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#289](https://github.com/adobe/spacecat-audit-worker/issues/289)) ([9d54932](https://github.com/adobe/spacecat-audit-worker/commit/9d549327c3f5e6e5ffc290028ad69f9ac4c8fecf))
+
+## [1.18.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.18.1...v1.18.2) (2024-06-13)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v1.29.0 ([#287](https://github.com/adobe/spacecat-audit-worker/issues/287)) ([eff7f8d](https://github.com/adobe/spacecat-audit-worker/commit/eff7f8d1e1ed8ccb9b5ddf4e240dac0b469648c9))
+
+## [1.18.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.18.0...v1.18.1) (2024-06-13)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v1.28.0 ([#283](https://github.com/adobe/spacecat-audit-worker/issues/283)) ([3d90431](https://github.com/adobe/spacecat-audit-worker/commit/3d90431b8abcbd14e1831a38bc31ba4057f43cca))
+
+# [1.18.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.17.2...v1.18.0) (2024-06-12)
+
+
+### Features
+
+* rum-api-client v2 with customer specific domain key support ([#282](https://github.com/adobe/spacecat-audit-worker/issues/282)) ([d34c6cf](https://github.com/adobe/spacecat-audit-worker/commit/d34c6cf75427d214aaf7e8fcebc11e4accbe2316))
+
 ## [1.17.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.17.1...v1.17.2) (2024-06-11)